feat: extract auth class; remove auth type
